requirements.txt
pytest
pyyaml
loguru
flask